refactor(codemod): move `query-cache` test cases to `default-import`
We want to cover the default, named, and namespaced imports in the
case of `QueryCache` as well.
