+++ codemods/v4/__tests__/key-transformation.test.js
@@ -29,10 +29,6 @@ defineTest(
   }
 )
 
-defineTest(__dirname, 'key-transformation', null, 'query-cache', {
-  parser: 'tsx',
-})
-
 defineTest(__dirname, 'key-transformation', null, 'type-arguments', {
   parser: 'tsx',
 })
